+++ CHANGELOG.md
@@ -9,6 +9,7 @@ You can also refer to commit logs to get details on what was implemented, fixed 
 
 - Move `secretInBody` and `customParameters` from `authConfig` to `clientConfig`, where they belong
 - Allow to override default UTF-8 encoding of the _Basic_ authorization header
+- Improvements to embedded authorization
 
 
 ### 3.0.1